move SpotDetectionMode from utils.py to _def.py
